requirements.txt
dj_database_url==2.2.0
Django==5.0.6
django_ckeditor==6.7.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
